flow UI update
Moderate update to the v.2.0 flow server UI based on artist mock-ups and Concord team meetings. This version still requires the features URL parameter to be set to 1.  Live connection to Pi was not available for testing so block evaluation, relay block, list of available pis, live recording to dataset file, and opening of dataset file all need to be tested. Further changes may be required to ensure proper functionality of these features.  Backwards compatiblity with previous UI versions has not been tested (nor was it requested by Concord staff).