make fixes based on classroom feedback and existing bugs: load dataset bug, invalid names, svg size
- check input fields (blocks, program name, dataset name, etc.) for invalid characters and leading/trailing space.  These changes are designed to prevent users from entering illegal values which cause trouble downstream.  Here are the rules that I have implemented:
When typing in an input field that expects a string, remove invalid characters (<>:"/\|?*').  This applies to the program name, the dataset name, and sequence names within a dataset.
When typing in an input field that expects a number, remove characters that are non-numeric and not a period ("." - for entering floats).
When focus leaves an input field that expects a string, remove lead and trailing white space.  This allows users to type string values with spaces such as "this is my dataset", but prevents invalid whitespace at start or end of input field when user is done typing.  This applies to the program name, the dataset name, and sequence names within a dataset.

- better handling of size of the SVG space.  This fixes a bug where pins and connections would not draw due to poorly sized SVG space.  This was mostly seen when the user scrolled down to see the lower portion of the diagram editor.

- add protection around opening invalid datasets.  Provide user with error message if opening null datasets or datasets with null metadata.  These cases were not being handled before and a silent exception was thrown.

- fix program save so we do NOT save sensor block values.  Previously we saving programs with a specific sensor value on a block.  When these programs were re-opened, the user could access states where these illegal, zombie sensor values would be displayed in the block.  So, for example, user might save a program with a temp sensor while connected to a Pi.  We would actually save the sensor value as well (so, for example, "30").  When the program was reopened, it was possible to re-display this invalid sensor value regardless of if we were not connected to a Pi, connected to a different Pi, etc.