Allow support of multiple sensors of same type
Allow support of multiple sensors of same type.  When generating a unique name for a sensor block, ensure that this name matches what is passed from the flow-client (which identifies and assigns names to sensors).  Also ensure that we properly update the nameHash object which maps block name passed from flow client to the actual flow blocks.  This object must be correctly maintained when loading a program, renaming a block, deleting a block, or undoing changes.
[#160838943]