disable rename or timer block
Disable rename of timer block to avoid conflict with sensor block names (which are used to determine which blocks receive which updates from the flow client).  Ideally we will turn this back on when we have a more sophisticated model of passing sensor values from the flow-client to the flow-server.
[#160838943]