160838943 multiple sensors same type (#106)
* use jslint to perform code refactor of program-editor-panel module

use jslint to perform code refactor of program-editor-panel module
[#160838943]

* Allow multiple sensors of same type

Allow multiple sensors of same type by generating the correct block name in getUniqueName() that matches the name created by the flow client and passed back to the flow server.
[#160838943]

* Allow support of multiple sensors of same type

Allow support of multiple sensors of same type.  When generating a unique name for a sensor block, ensure that this name matches what is passed from the flow-client (which identifies and assigns names to sensors).  Also ensure that we properly update the nameHash object which maps block name passed from flow client to the actual flow blocks.  This object must be correctly maintained when loading a program, renaming a block, deleting a block, or undoing changes.
[#160838943]

* disable rename or timer block

Disable rename of timer block to avoid conflict with sensor block names (which are used to determine which blocks receive which updates from the flow client).  Ideally we will turn this back on when we have a more sophisticated model of passing sensor values from the flow-client to the flow-server.
[#160838943]